+++ src/Exceptions/InvalidCallException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Qless\Exceptions;
+
+/**
+ * Qless\Exceptions\InvalidCallException
+ *
+ * @package Qless\Exceptions
+ */
+class InvalidCallException extends LogicException
+{
+}