+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- All qless exception now implements `Qless\Exceptions\ExceptionInterface`
 
 ## [2.1.0] - 2018-10-08
 ### Added